pyproject.toml
[project]
name = "lansweeperdataservice"
version = "0.1.0"
description = "Catalog-service microservice"
authors = [
    {name = "Yakubov Shakir",email = "Shakir.Yakubov@aton.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyodbc (>=5.2.0,<6.0.0)",
    "psycopg2 (==2.9.10)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "requests (>=2.32.4,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aioodbc (>=0.5.0,<0.6.0)",
    "alembic (>=1.16.4,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

